requirements/requirements.txt
email-validator==1.3.0
Flask==2.2.2
py3dns==3.2.1
python-dotenv==0.21.0
requests==2.28.1
slack-sdk==3.18.3
slackclient==2.9.4
